refactor: add auth rules file
