+++ src/auth/shield/rules/auth.rules.ts
@@ -1,0 +1,31 @@
+import { rule } from "graphql-shield";
+import { UNAUTHORIZED } from "../../../common/common.constants";
+import { Context } from "../../../context/context.service";
+import { getJti, getSub } from "../../auth.utils";
+
+export const isAuthenticated = rule({ cache: "contextual" })(
+  async (_parent, _args, { user }: Context) => {
+    if (!user) {
+      return UNAUTHORIZED;
+    }
+    return true;
+  }
+);
+
+export const hasValidRefreshToken = rule()(
+  async (
+    _parent,
+    _args,
+    {
+      claims: { refreshTokenClaims },
+      services: { refreshTokensService },
+    }: Context
+  ) => {
+    const jti = getJti(refreshTokenClaims);
+    const sub = getSub(refreshTokenClaims);
+    if (!jti || !sub) {
+      return UNAUTHORIZED;
+    }
+    return refreshTokensService.validateRefreshToken(jti, sub);
+  }
+);
