+++ src/auth/shield/shield.rules.ts
@@ -1,19 +1,8 @@
 import { rule } from "graphql-shield";
-import { UNAUTHORIZED } from "../../common/common.constants";
 import { Context } from "../../context/context.service";
 import { Event } from "../../events/models/event.model";
 import { GroupPrivacy } from "../../groups/group-configs/models/group-config.model";
-import { getJti, getSub } from "../auth.utils";
 import { hasNodes, hasServerPermission } from "./shield.utils";
-
-export const isAuthenticated = rule({ cache: "contextual" })(
-  async (_parent, _args, { user }: Context) => {
-    if (!user) {
-      return UNAUTHORIZED;
-    }
-    return true;
-  }
-);
 
 export const canCreateServerInvites = rule()(
   async (_parent, _args, { permissions }: Context) =>
@@ -43,24 +32,6 @@ export const canManageServerRoles = rule()(
 export const canRemoveMembers = rule()(
   async (_parent, _args, { permissions }: Context) =>
     hasServerPermission(permissions, "removeMembers")
-);
-
-export const hasValidRefreshToken = rule()(
-  async (
-    _parent,
-    _args,
-    {
-      claims: { refreshTokenClaims },
-      services: { refreshTokensService },
-    }: Context
-  ) => {
-    const jti = getJti(refreshTokenClaims);
-    const sub = getSub(refreshTokenClaims);
-    if (!jti || !sub) {
-      return UNAUTHORIZED;
-    }
-    return refreshTokensService.validateRefreshToken(jti, sub);
-  }
 );
 
 export const isUserInPublicPost = rule()(async (_parent, _args, _ctx, info) =>
